Cahnged attribute binding method
In order to be more flexible I have changed the attribute binding to be
only based upon variables, as follows:

<img data-bind_attribute="src" src="/images/${user.id}"/>

As the example shows, this allows embedding into the attribute when the
model does not hold all of the data (ie. does not have the url prefix,
in this case). It is also more succinct.

The previous method (defining a "data-bind_attribute_to") has been
removed to keep the source small and clean.